pytorch_flask/requirements.txt
Flask==2.1.1
Flask_Cors==3.0.10
grad_cam==1.3.7
numpy==1.21.2
opencv_python==4.5.5.64
Pillow==9.1.0
torch==1.10.1
torchvision==0.11.2
Werkzeug==2.1.1
